Move data loading functions to a new module
